Add documentation for x.py
Rust has switched over from old makefiles to the new build system that
uses x.py to bootstrap the repo and orchestrate building, testing, and
documenting all of the components for rustc. This system isn't as well
documented for newcomers who might not know how to use it. This adds
a skeleton with common use cases that can be expanded upon in the future
as the build system evolves or adding more documentation to current
functionality.