Improve infrastructure documentation
This adds some further documentation about the rust infrastructure
based on the description of the different projects in their respective
documentation.

Additionally it also removes the dead links to the outdated buildbot
(buildbot.rust-lang.org) and the crater overview which does not exist
in this form (crater.rust-lang.org).

Responsible maintainers to the different projects has not been added
or altered to prevent mistakes, since I'm not familiar with the internal
responsibilities of the involved parties.

The `https://rusty-dash.com/fcp` link also seems to point to a dead link,
but it has been kept because it might be a page which is only accessible
through an internal network (since it's reserved for team members).