update all dates in state-of-rust features table
I've gone though each of the tracking issues and set the date to roughly the latest update, in some cases I've just chosen the latest comment, on others I've marked it as the assignment of the [C-tracking-issue](https://github.com/rust-lang/rust/labels/C-tracking-issue) label (many of which were added by Mark Simulacrum on July 22nd 2017. I did explicitly ignore any of the link-back links below the last comment so in some cases (say the non-ASCII idents rust-lang/rust#28979) it is possibly out of date after the debate in June. Some of the latest comments dates may not count as 'updates'.

is it worth adding some indication as to when this list was last updated to the top-text, so these dates don't seem more stale than they are?