channel-layout: Add a channel layout document
As part of an effort to better document how `rustup` does what
it does, this is the beginning of a document explaining how Rust's
release channels are stored on `static.rust-lang.org`.  Further
explanations will be needed as time goes by, especially as `rustup`
proceeds with `profiles` or if we do end up adopting _TUF_ as
Brian's internals post suggested we might.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>