Add note to backport release notes
We've historically forgotten to do this a few times, so let's include it
in the documentation so that we don't forget in the future.

This also adjusts the remote we use to `origin` as that's what I have
locally (and is more common than rust-lang I suspect, given that it is
the git default).