Update username/github link of Gankra
Username previously listed for Gankra linked to a redirect github page.
This was noticed by @Moongoodgirl who mentioned it on Freenode/##rust.
I have also replaced the pronouns on the page with gender neutral
'they'. However, the ideal solution would be to get in touch with Alexis
and ask them what pronouns they would like on the page.

Reported-by: @Moongoodgirl
Signed-off-by: @aoeixsz4
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Your branch is up to date with 'origin/master'.
#
# Changes to be committed:
#	modified:   fott.md
#