Fixed outdated repository references
The Github repository references that describe which repositories have supported short linkifier notations were outdated for chalk, ena and polonius.
All three projects are no longer in rust-lang-nursery but in rust-lang. The resulting links will still work but might confuse new users.
The current linkifier settings for rust-lang.zulipchat.com already use the updated link for chalk but not for the other two.