Change “Release Date” to “Will be stable on”.
I believe the current heading “Release Date” is hard to understand; the first obvious interpretation is “The version currently in this channel was released to this channel on…”, which is falsifiable since non-stable channel dates are in the future, but it's still a speedbump. I believe that relabeling it to “Will be stable on” will make the table easier to understand by newcomers.