Don't use quotes around the Launching Pad team's name.
I'm not entirely certain what the RFC authors' intent was with this
style, but my guess was that it was because it was proposing a new term.
Since the team is now a real team, and this document is no longer being
framed as a proposal, this removes it and uses a consistent style.