Grant infra-admins permissions for repo changes
The `repo` directory manages both access as well as the configuration
and automation of repositories. Since all repositories are now managed
by the team repo and we are adding more and more automation, it makes
sense to give the infra-admins the authority to approve and merge these
configuration changes, while permission changes still need to be
approved by the team-repo-admins.