Add more platforms
This adds more platforms that are managed by the Rust project. These are
mostly just stubs. I would like to get them filled out more in the
future. The general idea is to record and track their existence and who
is responsible for them. The future intent is to extend these to include
the policies for how they get managed (or at least share with the X
content), and any other useful bits.