Remove old website config
This removes an old file that was used with an old system that we used
with AWS. This hasn't been used in a long time, and since then we've
also switched to just using GitHub Pages as-is.

There's more context in the discussion at
https://rust-lang.zulipchat.com/#narrow/channel/242791-t-infra/topic/forge.20cloudfront/with/535060344
and https://github.com/rust-lang/simpleinfra/issues/731.